Move write HTTP handler into own package
Move the write (ingest) handler to its own package.
