+++ main.go
@@ -1,28 +1,20 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"os"
-	"sort"
-	"sync"
 	"time"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/snappy"
 	v1API "github.com/mattbostock/athensdb/internal/api/v1"
 	"github.com/mattbostock/athensdb/internal/cluster"
-	"github.com/mattbostock/athensdb/internal/remote"
+	"github.com/mattbostock/athensdb/internal/write"
 	"github.com/prometheus/common/route"
-	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/storage/tsdb"
 	"golang.org/x/net/context"
-	"golang.org/x/net/context/ctxhttp"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
 
@@ -30,10 +22,7 @@ const (
 	defaultHTTPAddr = "localhost:9080"
 	defaultPeerAddr = "localhost:7946"
 
-	apiRoute   = "/api/v1"
-	writeRoute = "/receive"
-
-	httpHeaderInternalWrite = "X-AthensDB-Internal-Write-Version"
+	apiRoute = "/api/v1"
 )
 
 var (
@@ -121,108 +110,9 @@ func main() {
 	var api = v1API.NewAPI(queryEngine, localStorage)
 	api.Register(router.WithPrefix(apiRoute))
 
-	router.Post(writeRoute, func(w http.ResponseWriter, r *http.Request) {
-		compressed, err := ioutil.ReadAll(r.Body)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		reqBuf, err := snappy.Decode(nil, compressed)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		var req remote.WriteRequest
-		if err := proto.Unmarshal(reqBuf, &req); err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		appender, err := localStorage.Appender()
-		if err != nil {
-			// FIXME: Make error more useful
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-		defer appender.Commit()
-
-		var samples uint32
-		for _, ts := range req.Timeseries {
-			m := make(labels.Labels, 0, len(ts.Labels))
-			for _, l := range ts.Labels {
-				m = append(m, labels.Label{
-					Name:  l.Name,
-					Value: l.Value,
-				})
-			}
-			sort.Sort(m)
-
-			for _, s := range ts.Samples {
-				// FIXME: Look at using AddFast
-				appender.Add(m, s.TimestampMs, s.Value)
-				samples++
-			}
-		}
-
-		// This is an internal write, so don't replicate it to other nodes
-		if r.Header.Get(httpHeaderInternalWrite) != "" {
-			log.Debugf("Received %d samples from another node in the cluster", samples)
-			return
-		}
-
-		var wg sync.WaitGroup
-		// FIXME: Avoid panic if the cluster is not yet initialised
-		nodes := cluster.Nodes()
-		var wgErrChan = make(chan error, len(nodes))
-		for _, node := range nodes {
-			if node.Name() == cluster.LocalNode().Name() {
-				log.Debugf("Skipping local node %s", node)
-				continue
-			}
-
-			wg.Add(1)
-			go func(n *cluster.Node) {
-				defer wg.Done()
-
-				log.Debugf("Writing %d samples to %s", samples, n)
-
-				httpAddr, err := n.HTTPAddr()
-				if err != nil {
-					wgErrChan <- err
-					return
-				}
-				apiURL := fmt.Sprintf("%s%s%s", "http://", httpAddr, writeRoute)
-				nodeReq, err := http.NewRequest("POST", apiURL, bytes.NewBuffer(compressed))
-				if err != nil {
-					wgErrChan <- err
-					return
-				}
-				nodeReq.Header.Add("Content-Encoding", "snappy")
-				nodeReq.Header.Set("Content-Type", "application/x-protobuf")
-				nodeReq.Header.Set("X-Prometheus-Remote-Write-Version", "0.1.0")
-				nodeReq.Header.Set(httpHeaderInternalWrite, "0.0.1")
-
-				// FIXME set timeout using context
-				httpResp, err := ctxhttp.Do(context.TODO(), http.DefaultClient, nodeReq)
-				if err != nil {
-					wgErrChan <- err
-					return
-				}
-				defer httpResp.Body.Close()
-			}(node)
-		}
-		// FIXME cancel requests if one fails
-		wg.Wait()
-
-		select {
-		case err := <-wgErrChan:
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		default:
-		}
-	})
+	write.SetLogger(log.StandardLogger())
+	write.SetStore(localStorage)
+	router.Post(write.Route, write.Handler)
 
 	if err := cluster.Join(&cluster.Config{
 		HTTPAdvertiseAddr: *config.httpAdvertiseAddr,
